resource-custodian/resource-custodian/requirements.txt
azure-mgmt-resource==2.1.0
azure-monitor==0.3.1
